requirements.txt
beautifulsoup4==4.13.3
PyYAML==6.0.2
requests==2.32.3